bicycle-service/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
asyncpg>=0.29.0
aio-pika>=9.4.1
python-jose[cryptography]>=3.3.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
pytest>=8.2.0
pytest-asyncio>=0.23.6
httpx>=0.27.0
python-dotenv>=1.0.1
cryptography>=42.0.0
